@@ -1,7 +1,8 @@
 {
-    "name": "atukai/dashboard-grid",
-    "description": "Rapid admin panel builder",
-    "homepage": "https://github.com/atukai/dashboard",
+    "name": "atukai/gridly",
+    "license": "MIT",
+    "description": "Gridly - PHP Grid Manager",
+    "homepage": "https://github.com/atukai/gridly",
     "authors": [
         {
             "name": "Alexander Tukai",
@@ -9,22 +10,32 @@
         }
     ],
     "require": {
-        "php": "^7.1",
-        "zendframework/zend-paginator": "^2.8",
-        "pagerfanta/pagerfanta": "^1.0"
+        "php": "^8.1",
+        "ext-pdo": "*",
+        "psr/http-message": "^1.0.1",
+        "symfony/config": "^6.1",
+        "symfony/console": "^6.1",
+        "symfony/yaml": "^6.1"
     },
     "require-dev": {
-        "phpunit/phpunit": "*",
-        "squizlabs/php_codesniffer": "*"
+        "doctrine/collections": "^1.8",
+        "doctrine/dbal": "^3.4",
+        "doctrine/orm": "^2.13",
+        "laminas/laminas-db": "^2.16",
+        "laminas/laminas-paginator": "^2.17",
+        "pagerfanta/pagerfanta": "^3.7",
+        "phpunit/phpunit": "^9.5",
+        "roave/security-advisories": "dev-latest",
+        "twig/twig": "^3.4.0"
     },
     "autoload": {
         "psr-4": {
-            "Dashboard\\": "src/"
+            "Gridly\\": "src/"
         }
     },
     "autoload-dev": {
         "psr-4": {
-            "DashboardTest\\": "tests/Dashboard"
+            "Gridly\\Test\\": "tests/"
         }
     }
 }
