@@ -1,55 +1,88 @@
 <?php
 
-namespace Dashboard;
+namespace Gridly;
 
-class ResultSet implements \Countable
+use ArrayIterator;
+use Countable;
+use Gridly\Column\Column;
+use Gridly\Column\Definitions;
+use Gridly\Column\Exception;
+use Gridly\Row\Row;
+use IteratorAggregate;
+
+class ResultSet implements Countable, IteratorAggregate
 {
-    /** @var iterable  */
-    private $rawData;
-
-    /** @var array  */
-    private $rows = [];
-
+    private Row $headersRow;
+    
+    /** @var Row[] */
+    private array $rows;
+    
     /**
-     * @param iterable $data
-     * @param array $hiddenColumns
+     * @throws Exception
      */
-    public function __construct(iterable $data, array $hiddenColumns = [])
+    public function __construct(iterable $data, Definitions $columnDefinitions, array $columnDecorators = [])
     {
-        $this->rawData = $data;
-        $this->processData($data, $hiddenColumns);
+        $this->rows = [];
+        $this->prepareRows($data, $columnDefinitions, $columnDecorators);
     }
-
-    /**
-     * @return array
-     */
-    public function getRows()
+    
+    public function getHeadersRow(): Row
+    {
+        return $this->headersRow;
+    }
+    
+    public function getRows(): array
     {
         return $this->rows;
     }
-
-    /**
-     * @return int
-     */
-    public function count()
+    
+    public function count(): int
     {
         return count($this->rows);
     }
 
-    /**
-     * @param iterable $data
-     * @param array $hiddenColumns
-     */
-    private function processData(iterable $data, array $hiddenColumns = []): void
+    public function getIterator(): ArrayIterator
     {
-        foreach ($data as $entry) {
-            $row = [];
+        return new ArrayIterator($this->rows);
+    }
+    
+    /**
+     * @throws Exception
+     */
+    private function prepareRows(iterable $entries, Definitions $columnDefinitions, array $columnDecorators = []): void
+    {
+        foreach ($entries as $i => $entry) {
+            $row = new Row();
+            
             foreach ($entry as $name => $value) {
-                if (!in_array($name, $hiddenColumns)) {
-                    $row[] = new Column($name, $value);
+                $columnDefinition = $columnDefinitions->get($name);
+                
+                if (!$columnDefinition->isVisible()) {
+                    continue;
                 }
-            }
+                
+                $column = new Column(
+                    $name,
+                    $value,
+                    $columnDefinition->getType(),
+                    $columnDefinition->getLabel(),
+                    $columnDefinition->isSortable(),
+                    $columnDefinition->isFilterable()
+                );
+                
+                if (isset($columnDecorators[$column->name()])) {
+                    $decorator = $columnDecorators[$column->name()];
+                    $column = $decorator($column, $entry);
+                }
 
+                $row->addColumn($column);
+            }
+            
+            // First row for headers
+            if ($i === 0) {
+                $this->headersRow = $row;
+            }
+            
             $this->rows[] = $row;
         }
     }
